improve staging performance a bit by reducing excess build deps
in the content snap job we create a dump of stuff we explicitly put
into the dev tarball, we'll substract those known staged packages from the
dev packages to reduce the amount of needlessly staged packages (needless
because we alredy have them in the tarball stage, so we dont need them
again)

FTR: they can still be installed if they are deps of another thing we have
a build dependency one. e.g. libkdegames deps kf5 so the dependees will
still be installed as build-depends even though not explicitly listed