fix source hanlding [by importing debian classes from pangea-tooling]
apt-cache showsrc has fairly daft behavior. It lists all possible versions
AND it does so in an unordered fashion. As a result we may get a kde4
version or a super out of date kf5 version. Both are not useful and
defeat the purpose of the feature.

To not overly tie any of this to pangea's infrastructure and job structure
there's now a copy of the debian classes so we can parse control files.
Instead of showsrcing we now apt-get source, and then parse the control
file of the source. This allows higher quality parsing and "fixes" the
problem of using showsrc and its unreliable behavior.

Caveat: since pangea has no actual build dep resolver class but outsources
it we need to do some minor naughty juggling of deps. In alternates the
first always wins. Architecture and version restrictions are ripped out
entirely. Should be good for most things, but who knows.