try my luck at preventing excess runtime crap from ending in the snap
the frameworks snap now tries to build a more comprehensive list of both
stage's content. for runtime we now have a dedicated part which will
currently attempt to stage only qml dependencies as determined via
the deb packaging.
this part uses special rigging which will explicitly exclude specified
debs from being staged here.

i.e. what happens is that we will tell apt to download everything, but then
only extract the debs which were not otherwise blacklisted.
this means that by using the (comprehensive) stage content from the
frameworks build, we should be able to accurately prevent everything that
is in the content snap from being staged again for the app snap

(in theory at least)